--- nd4j-jcublas-parent/nd4j-jcublas-common/src/test/resources/log4j.properties
@@ -1,28 +1,0 @@
-
-#
-# Copyright 2015 Skymind,Inc.
-#
-#    Licensed under the Apache License, Version 2.0 (the "License");
-#    you may not use this file except in compliance with the License.
-#    You may obtain a copy of the License at
-#
-#        http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS,
-#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#    See the License for the specific language governing permissions and
-#    limitations under the License.
-#
-
-log4j.rootLogger=ERROR, Console
-log4j.appender.Console=org.apache.log4j.ConsoleAppender
-log4j.appender.Console.layout=org.apache.log4j.PatternLayout
-log4j.appender.Console.layout.ConversionPattern=%d{ABSOLUTE} %-5p ~ %m%n
-
-log4j.appender.org.springframework=DEBUG
-log4j.appender.org.deeplearning4j=INFO
-
-log4j.logger.org.springframework=INFO
-log4j.logger.org.deeplearning4j=INFO
-
