--- nd4j-jcublas-parent/nd4j-jcublas-common/src/test/resources/logback.xml
@@ -1,46 +1,0 @@
-<!--
-  ~ Copyright 2015 Skymind,Inc.
-  ~
-  ~    Licensed under the Apache License, Version 2.0 (the "License");
-  ~    you may not use this file except in compliance with the License.
-  ~    You may obtain a copy of the License at
-  ~
-  ~        http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~    Unless required by applicable law or agreed to in writing, software
-  ~    distributed under the License is distributed on an "AS IS" BASIS,
-  ~    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~    See the License for the specific language governing permissions and
-  ~    limitations under the License.
-  -->
-
-<configuration>
-
-	
-
-	<appender name="FILE" class="ch.qos.logback.core.FileAppender">
-		<file>logs/application.log</file>
-		<encoder>
-			<pattern>%date - [%level] - from %logger in %thread
-				%n%message%n%xException%n</pattern>
-		</encoder>
-	</appender>
-
-	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-		<encoder>
-			<pattern> %logger{15} - %message%n%xException{5}
-			</pattern>
-		</encoder>
-	</appender>
-    <logger name="org.apache.catalina.core" level="DEBUG" />
-	<logger name="org.springframework" level="DEBUG" />
-	<logger name="org.deeplearning4j" level="INFO" />
-    <logger name="org.nd4j" level="INFO" />
-
-
-	<root level="ERROR">
-		<appender-ref ref="STDOUT" />
-		<appender-ref ref="FILE" />
-	</root>
-
-</configuration>
